requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-model-utils==4.3.1
django-notifications-hq==1.8.2
django-resized==1.0.2
idna==3.4
jsonfield==3.1.0
paynow==1.0.4
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
typing_extensions==4.7.0
urllib3==2.0.3
